requirements.txt
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
langchain>=0.0.267
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
streamlit>=1.25.0
peft>=0.4.0
trl>=0.7.2
datasets>=2.12.0
pypdf>=3.15.1
docx2txt>=0.8
tqdm>=4.65.0
PyYAML>=6.0
scikit-learn>=1.2.2
huggingface_hub>=0.16.4
gradio>=3.40.1
numpy>=1.24.3
pandas>=2.0.2
matplotlib>=3.7.1
wandb>=0.15.8
einops>=0.6.1
